Pi 5/stt/requirements.txt
argcomplete==3.6.0
attrs==25.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorlog==6.9.0
dependency-groups==1.3.0
distlib==0.3.9
filelock==3.17.0
idna==3.10
mypy-extensions==1.0.0
nox==2025.2.9
packaging==24.2
paho-mqtt==2.1.0
parse==1.20.2
platformdirs==4.3.6
PyAudio==0.2.14
pycparser==2.22
requests==2.32.3
sounddevice==0.5.1
srt==3.5.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
virtualenv==20.29.3
vosk==0.3.45
vox==0.0.1
websockets==15.0.1
